fix: replace signup nickname page
